+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 
+- Introduced more informative message about compression result: original file size, compressed file size and compression percentage.
 - Added "Compress (Save As)" action. Mostly to provide a solution for the case when the current file's folder is not writable.
 - Disabled "Compress" action when current folder is not writable.
 - Use balloon to show success and error notifications instead of dialog.